// src/types/socialProfile.ts
export type SocialProfile = {
  fb_user_id: string;
  page_id: string;
  company_id: number;
  page_name: string;
  instagram_business_account: string;
  instagram_username: string;
  active: number;
};

export type SocialProfilePages = {
  instagram_business_account: boolean;
  instagram_username: string;
  page_id: string;
  page_name: string;
};